pyproject.toml
[project]
name = "unihack2026"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "celery[redis]>=5.6.2",
    "flask>=3.1.3",
    "flask-security-too>=5.7.1",
    "flask-sqlalchemy>=3.1.1",
    "pydantic>=2.12.5",
    "pyopenssl>=25.3.0",
    "pytz>=2026.1.post1",
    "requests>=2.32.5",
    "waitress>=3.0.2",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "ruff>=0.15.6",
]

[tool.uv.workspace]
members = [
    "bootstrap-icons",
]
